+++ dev/tests/spec/create.spec.js
@@ -27,6 +27,7 @@ describe('crossroads.create()', function(){
             expect( t1 ).toBe( 'lorem_ipsum' );
         });
 
+
         it('shouldn\'t affect static instance', function(){
             var t1;
             var cr = crossroads.create();
@@ -42,6 +43,7 @@ describe('crossroads.create()', function(){
             expect( t1 ).toBe( 'lorem_ipsum' );
         });
 
+
         it('shouldn\'t be affected by static instance', function(){
             var t1;
             var cr = crossroads.create();
@@ -56,6 +58,38 @@ describe('crossroads.create()', function(){
 
             expect( t1 ).toBe( 'lorem_ipsum' );
         });
+
+
+        it('should allow a different lexer per router', function () {
+            var cr = crossroads.create();
+            var count = 0;
+            cr.patternLexer = {
+                getParamIds : function(){
+                    return ['a','b'];
+                },
+                getOptionalParamsIds : function(){
+                    return [];
+                },
+                getParamValues : function(){
+                    return [123, 456];
+                },
+                compilePattern : function(){
+                    return (/foo-bar/);
+                }
+            };
+            var vals = [];
+            var inc = function(a, b){
+                vals[0] = a;
+                vals[1] = b;
+                count++;
+            };
+            cr.addRoute('test', inc);
+            cr.parse('foo-bar');
+            expect( count ).toEqual( 1 );
+            expect( vals ).toEqual( [123, 456] );
+            expect( cr.patternLexer ).not.toBe( crossroads.patternLexer );
+        });
+
 
     });