+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Crossroads.js Changelog #
 
+## Next ##
+
+ - improve `Route.interpolate()` to support query strings. (#76)
+ - make it possible to have a different patternLexer per router. (#67)
+ - add trailing/leading ";" to crossroads.min.js to avoid concat issues. (#73)
+
+
 ## v0.11.0 (2012/10/31) ##
 
 ### API Changes